move event widget to own component file
